Organized packages to a more understandable way.
