+++ src/game/ai/worm/LightWorm.java
@@ -1,4 +1,4 @@
-package game.ui;
+package game.ai.worm;
 
 import javafx.scene.image.Image;
 
@@ -6,7 +6,7 @@ public class LightWorm extends Worm{
 	private static Image wormyRight = new Image("/game/sprites/worm/Wormy_Right.png");
 	private static Image wormyLeft = new Image("/game/sprites/worm/Wormy_Left.png");
 	
-	LightWorm(double x, double y, int spriteWidth, int spriteHeight){
+	public LightWorm(double x, double y, int spriteWidth, int spriteHeight){
 		super(x, y, spriteWidth, spriteHeight, wormyRight, wormyLeft, 5, 5, 5);
 	}
 }
