+++ src/game/ai/worm/RangedWorm.java
@@ -1,4 +1,4 @@
-package game.ui;
+package game.ai.worm;
 
 public class RangedWorm {
