+++ src/game/ui/GameScreen.java
@@ -1,0 +1,7 @@
+package game.ui;
+
+import javafx.scene.layout.Pane;
+
+public class GameScreen extends Pane{
+
+}
